--- src/micm/micm_c_api.cpp
@@ -1,28 +1,0 @@
-#include <musica/micm.hpp>
-#include <musica/micm_c.h>
-
-#include <iostream>
-
-void create_micm(void **micm)
-{
-    *micm = new MICM();
-}
-
-void delete_micm(void **micm)
-{
-    if (*micm)
-    {
-        delete static_cast<MICM *>(*micm);
-        *micm = nullptr;
-    }
-}
-
-int micm_create_solver(void **micm, const char *config_path)
-{
-    return static_cast<MICM *>(*micm)->create_solver(std::string(config_path));
-}
-
-void micm_solve(void **micm, double time_step, double temperature, double pressure, int num_concentrations, double *concentrations)
-{
-    static_cast<MICM *>(*micm)->solve(time_step, temperature, pressure, num_concentrations, concentrations);
-}
